requirements.txt
beautifulsoup4==4.9.3
boto3==1.16.18
botocore==1.19.18
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
future==0.18.2
idna==2.10
jmespath==0.10.0
probableparsing==0.0.1
PyPDF2==1.26.0
python-crfsuite==0.9.7
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.0
s3transfer==0.3.3
six==1.15.0
soupsieve==2.0.1
urllib3==1.26.2
usaddress==0.5.10
